+++ hiperion_web/src/main/java/ec/com/avila/emision/web/domain/CondicionEspecial.java
@@ -13,7 +13,7 @@ package ec.com.avila.emision.web.domain;
  */
 public class CondicionEspecial {
 
-	private Integer id;
+	private Long id;
 	private String nombre;
 
 	/**
@@ -21,7 +21,7 @@ public class CondicionEspecial {
 	 * @param idAnexo
 	 * @param nombreDetalleAnexo
 	 */
-	public CondicionEspecial(Integer id, String nombre) {
+	public CondicionEspecial(Long id, String nombre) {
 		this.id = id;
 		this.nombre = nombre;
 	}
@@ -29,7 +29,7 @@ public class CondicionEspecial {
 	/**
 	 * @return the id
 	 */
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
 
@@ -37,7 +37,7 @@ public class CondicionEspecial {
 	 * @param id
 	 *            the id to set
 	 */
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
